Add HAI demo, the human anatomy demo to demonstrate LOD reader
The demo takes as input an xml file (VTK generated) containing blocks(pieces) of
pointers to different LOD polydata(.vtp).
Only 1 Level of Detail (LOD) of a piece is used at a time.
The right panel lists the pieces and their LOD of the main xml file.
The user can interact with each piece to request a particular LOD
(not implemented yet).